backend/pyproject.toml
[project]
name = "apex-backend"
version = "0.1.0"
description = "FastAPI HTTP + SSE server for the Apex Agent."
requires-python = ">=3.11"
dependencies = [
    "apex-agent-core",
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "argon2-cffi>=25.1.0",
    "slowapi>=0.1.9",
    "sse-starlette>=3.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apex_server"]
